backend/requirements.txt
Django==5.2.6
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
pandas==2.3.2
pyotp==2.9.0
qrcode[pil]==8.2